refactor default table formatter into pkg/app/formatter/table
