+++ pkg/app/context_ls.go
@@ -15,18 +15,15 @@
 package app
 
 import (
-	"fmt"
-	"io"
 	"os"
 
 	"github.com/cockroachdb/errors"
-	"github.com/olekukonko/tablewriter"
 	"github.com/urfave/cli/v2"
 
 	"github.com/tensorchord/envd/pkg/app/formatter"
 	"github.com/tensorchord/envd/pkg/app/formatter/json"
+	"github.com/tensorchord/envd/pkg/app/formatter/table"
 	"github.com/tensorchord/envd/pkg/home"
-	"github.com/tensorchord/envd/pkg/types"
 )
 
 var CommandContextList = &cli.Command{
@@ -46,43 +43,9 @@ func contextList(clicontext *cli.Context) error {
 	format := clicontext.String("format")
 	switch format {
 	case "table":
-		renderContext(os.Stdout, contexts)
+		table.RenderContext(os.Stdout, contexts)
 	case "json":
 		return json.PrintContext(contexts)
 	}
 	return nil
-}
-
-func renderContext(w io.Writer, contexts types.EnvdContext) {
-	table := tablewriter.NewWriter(w)
-	table.SetHeader([]string{"context", "builder", "builder addr", "runner", "runner addr"})
-
-	table.SetAutoWrapText(false)
-	table.SetAutoFormatHeaders(true)
-	table.SetHeaderAlignment(tablewriter.ALIGN_LEFT)
-	table.SetAlignment(tablewriter.ALIGN_LEFT)
-	table.SetCenterSeparator("")
-	table.SetColumnSeparator("")
-	table.SetRowSeparator("")
-	table.SetHeaderLine(false)
-	table.SetBorder(false)
-	table.SetTablePadding("\t") // pad with tabs
-	table.SetNoWhiteSpace(true)
-
-	for _, p := range contexts.Contexts {
-		envRow := make([]string, 5)
-		if p.Name == contexts.Current {
-			envRow[0] = fmt.Sprintf("%s (current)", p.Name)
-		} else {
-			envRow[0] = p.Name
-		}
-		envRow[1] = string(p.Builder)
-		envRow[2] = fmt.Sprintf("%s://%s", p.Builder, p.BuilderAddress)
-		envRow[3] = string(p.Runner)
-		if p.RunnerAddress != nil {
-			envRow[4] = stringOrNone(*p.RunnerAddress)
-		}
-		table.Append(envRow)
-	}
-	table.Render()
 }
